% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shewhart.R
\name{Gompertz}
\alias{Gompertz}
\title{Gompertz Function}
\usage{
Gompertz(x, y0, ymax, k, lag)
}
\arguments{
\item{x}{A numeric vector (the x-axis).}

\item{y0}{Numeric. The initial value of the curve.}

\item{ymax}{Numeric. The asymptotic maximum value of the curve.}

\item{k}{Numeric. The growth rate parameter.}

\item{lag}{Numeric. The horizontal shift (lag).}
}
\value{
A numeric vector of length \code{length(x)}, with the Gompertz function
  values for each element of \code{x}.
}
\description{
Computes the value of the Gompertz function parameterized by \code{(y0, ymax, k, lag)}.
}
\details{
The Gompertz function is often used in growth curve modeling, where:
\deqn{ G(x) = y0 + (ymax - y0)\exp\Bigl[-\exp\bigl(k(\mathrm{lag}-x)/(ymax-y0)+1\bigr)\Bigr]. }
}
\examples{
x_seq <- seq(0, 20, by = 0.5)
y_gomp <- Gompertz(x_seq, y0 = 1, ymax = 10, k = 0.3, lag = 10)
plot(x_seq, y_gomp, type = "l", main = "Gompertz Curve")

}
