% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shewhart.R
\name{loglog}
\alias{loglog}
\title{Log-log transformation}
\usage{
loglog(x, alpha = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{alpha}{A numeric parameter that is subtracted before taking the log.
Default is \code{1}.}
}
\value{
A numeric vector with the log-log transformation applied.
}
\description{
Applies the transformation:
\deqn{ \log\!\bigl(\log\bigl(\tfrac{x}{\alpha} + 1\bigr) + 1\bigr). }
}
\details{
Useful in certain scenarios where a "double log" scale is appropriate.
}
\examples{
x <- 1:10
transformed <- loglog(x)
transformed

}
