% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{shewhart}
\alias{shewhart}
\title{Get Shewhard phases using 7 points rule}
\usage{
shewhart(
  data,
  values_col,
  index_col,
  start_base = 10,
  type = "ggplot",
  rule = "7 points",
  method = "log",
  locale = "pt_BR.UTF-8",
  ...
)
}
\arguments{
\item{data}{A tibble white de}

\item{values_col}{Column with values to analyze}

\item{index_col}{Column with sequence of dates or intergers}

\item{start_base}{Number of points in start of the series to use as base}

\item{phase_changes}{Vector with dates por positions for changes of phase}

\item{model}{One of log (default) and Gompetz.}
}
\description{
Get Shewhard phases using 7 points rule
}
